Implement in asm.js
Shaves off 200ms my test case, was it worth it?

Probably, though non-asm.js firefox is hella slow for
whatever reason (13 seconds). Chromium is at around ~1s
which is acceptable.

Now to put it in a web worker to avoid blocking.

Also looking at ways to get incremental results out of
the algorithm, but it doesn't look very possible unfortunately.